+++ charmcraft.yaml
@@ -17,5 +17,4 @@ parts:
       - rustc
       - cargo
     charm-binary-python-packages:
-      - jsonschema
-      - psycopg2-binary==2.9.3
+      - psycopg2-binary==2.9.5
