+++ metadata.yaml
@@ -35,4 +35,4 @@ requires:
 storage:
   pgdata:
     type: filesystem
-    location: /var/lib/postgresql/data
+    location: /var/snap/charmed-postgresql/common/postgresql/pgdata
