+++ tox.ini
@@ -75,13 +75,7 @@ deps =
     psycopg2-binary
     -r {tox_root}/requirements.txt
 commands =
-    # Download patroni resource to use in the charm deployment.
-    sh -c 'stat patroni.tar.gz > /dev/null 2>&1 || curl "https://github.com/zalando/patroni/archive/refs/tags/v2.1.4.tar.gz" -L -s > patroni.tar.gz'
     pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/test_backups.py
-    # Remove the downloaded resource.
-    sh -c 'rm -f patroni.tar.gz'
-allowlist_externals =
-    sh
 
 [testenv:charm-integration]
 description = Run charm integration tests
@@ -90,15 +84,9 @@ pass_env =
     CI
     CI_PACKED_CHARMS
 commands =
-    # Download patroni resource to use in the charm deployment.
-    sh -c 'stat patroni.tar.gz > /dev/null 2>&1 || curl "https://github.com/zalando/patroni/archive/refs/tags/v2.1.4.tar.gz" -L -s > patroni.tar.gz'
     poetry install --with integration
     poetry export -f requirements.txt -o requirements.txt
     poetry run pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/test_charm.py
-    # Remove the downloaded resource.
-    sh -c 'rm -f patroni.tar.gz'
-allowlist_externals =
-    sh
 
 [testenv:database-relation-integration]
 description = Run database relation integration tests
@@ -107,16 +95,10 @@ pass_env =
     CI
     CI_PACKED_CHARMS
 commands =
-    # Download patroni resource to use in the charm deployment.
-    sh -c 'stat patroni.tar.gz > /dev/null 2>&1 || curl "https://github.com/zalando/patroni/archive/refs/tags/v2.1.4.tar.gz" -L -s > patroni.tar.gz'
     poetry install --with integration
     poetry export -f requirements.txt -o requirements.txt
     poetry export --only relation_charm -f requirements.txt -o tests/integration/new_relations/application-charm/requirements.txt
     poetry run pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/new_relations/test_new_relations.py
-    # Remove the downloaded resource.
-    sh -c 'rm -f patroni.tar.gz'
-allowlist_externals =
-    sh
 
 [testenv:db-relation-integration]
 description = Run db relation integration tests
@@ -125,15 +107,9 @@ pass_env =
     CI
     CI_PACKED_CHARMS
 commands =
-    # Download patroni resource to use in the charm deployment.
-    sh -c 'stat patroni.tar.gz > /dev/null 2>&1 || curl "https://github.com/zalando/patroni/archive/refs/tags/v2.1.4.tar.gz" -L -s > patroni.tar.gz'
     poetry install --with integration
     poetry export -f requirements.txt -o requirements.txt
     poetry run pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/test_db.py
-    # Remove the downloaded resource.
-    sh -c 'rm -f patroni.tar.gz'
-allowlist_externals =
-    sh
 
 [testenv:db-admin-relation-integration]
 description = Run db-admin relation integration tests
@@ -142,15 +118,9 @@ pass_env =
     CI
     CI_PACKED_CHARMS
 commands =
-    # Download patroni resource to use in the charm deployment.
-    sh -c 'stat patroni.tar.gz > /dev/null 2>&1 || curl "https://github.com/zalando/patroni/archive/refs/tags/v2.1.4.tar.gz" -L -s > patroni.tar.gz'
     poetry install --with integration
     poetry export -f requirements.txt -o requirements.txt
     poetry run pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/test_db_admin.py
-    # Remove the downloaded resource.
-    sh -c 'rm -f patroni.tar.gz'
-allowlist_externals =
-    sh
 
 [testenv:ha-self-healing-integration]
 description = Run high availability self healing integration tests
@@ -159,15 +129,9 @@ pass_env =
     CI
     CI_PACKED_CHARMS
 commands =
-    # Download patroni resource to use in the charm deployment.
-    sh -c 'stat patroni.tar.gz > /dev/null 2>&1 || curl "https://github.com/zalando/patroni/archive/refs/tags/v2.1.4.tar.gz" -L -s > patroni.tar.gz'
     poetry install --with integration
     poetry export -f requirements.txt -o requirements.txt
     poetry run pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/ha_tests/test_self_healing.py
-    # Remove the downloaded resource.
-    sh -c 'rm -f patroni.tar.gz'
-allowlist_externals =
-    sh
 
 [testenv:password-rotation-integration]
 description = Run password rotation integration tests
@@ -176,15 +140,9 @@ pass_env =
     CI
     CI_PACKED_CHARMS
 commands =
-    # Download patroni resource to use in the charm deployment.
-    sh -c 'stat patroni.tar.gz > /dev/null 2>&1 || curl "https://github.com/zalando/patroni/archive/refs/tags/v2.1.4.tar.gz" -L -s > patroni.tar.gz'
     poetry install --with integration
     poetry export -f requirements.txt -o requirements.txt
     poetry run pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/test_password_rotation.py
-    # Remove the downloaded resource.
-    sh -c 'rm -f patroni.tar.gz'
-allowlist_externals =
-    sh
 
 [testenv:tls-integration]
 description = Run TLS integration tests
@@ -193,16 +151,10 @@ pass_env =
     CI
     CI_PACKED_CHARMS
 commands =
-    # Download patroni resource to use in the charm deployment.
-    sh -c 'stat patroni.tar.gz > /dev/null 2>&1 || curl "https://github.com/zalando/patroni/archive/refs/tags/v2.1.4.tar.gz" -L -s > patroni.tar.gz'
     poetry install --with integration
     poetry export -f requirements.txt -o requirements.txt
     poetry export --only ha_charm -f requirements.txt -o tests/integration/ha_tests/application-charm/requirements.txt
     poetry run pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/test_tls.py
-    # Remove the downloaded resource.
-    sh -c 'rm -f patroni.tar.gz'
-allowlist_externals =
-    sh
 
 [testenv:integration]
 description = Run all integration tests
@@ -211,17 +163,11 @@ pass_env =
     CI
     CI_PACKED_CHARMS
 commands =
-    # Download patroni resource to use in the charm deployment.
-    sh -c 'stat patroni.tar.gz > /dev/null 2>&1 || curl "https://github.com/zalando/patroni/archive/refs/tags/v2.1.4.tar.gz" -L -s > patroni.tar.gz'
     poetry install --with integration
     poetry export -f requirements.txt -o requirements.txt
     poetry export --only ha_charm -f requirements.txt -o tests/integration/ha_tests/application-charm/requirements.txt
     poetry export --only relation_charm -f requirements.txt -o tests/integration/new_relations/application-charm/requirements.txt
     poetry run pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/
-    # Remove the downloaded resource.
-    sh -c 'rm -f patroni.tar.gz'
-allowlist_externals =
-    sh
 
 [testenv:poetry-lock]
 description = Install, lock and export poetry dependencies
